refactor: move `message/packet/codec` to internal package
