+++ handler.go
@@ -28,10 +28,10 @@ import (
 	"reflect"
 	"time"
 
-	"github.com/lonnng/nano/codec"
+	"github.com/lonnng/nano/internal/codec"
 	"github.com/lonnng/nano/component"
-	"github.com/lonnng/nano/message"
-	"github.com/lonnng/nano/packet"
+	"github.com/lonnng/nano/internal/message"
+	"github.com/lonnng/nano/internal/packet"
 	"github.com/lonnng/nano/session"
 )
