+++ handler_test.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/golang/protobuf/proto"
 	"github.com/lonnng/nano/component"
-	"github.com/lonnng/nano/message"
+	"github.com/lonnng/nano/internal/message"
 	"github.com/lonnng/nano/serialize/json"
 	"github.com/lonnng/nano/serialize/protobuf"
 	"github.com/lonnng/nano/session"
